1) removed progbar/add tffunction 2) implement mixed precision 3) reshuffle for small dataset 4) min_sum in d_train::forward()

+++ src/models/multiscale.py
@@ -30,7 +30,7 @@ class MSUnet():
                  dropout=0, 
                  normalize=True,
                  up_type='upsampling', 
-                 dtype='float32'):
+                 dtype='float16'):
 
 
         self.filters = filters
